+++ ProceduralPlanet/src/Window.h
@@ -19,7 +19,7 @@ private:
 public:
 	Window();
 	Window(int height, int width);
-	float getCurrentTime() { return glfwGetTime(); };
+	float getCurrentTime() { return (float)glfwGetTime(); };
 	Input* getInputManger();
 	float getAspectRatio() { return aspectRatio; };
 	void clearColor();
